Smart Menu 2017/src/superorder-master/requirements.txt
Django==1.10.6
Unipath==1.1
Pillow==4.3.0
requests==2.18.4
